requirements.txt
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
cupy-cuda12x>=13.0.0
transformers>=4.40.0
torchvision>=0.15.0
matplotlib>=3.7.0
datasets>=2.14.0
